app/src/test/java/com/kmvdata/kotlin/beatbox/SoundViewModelTest.kt
package com.kmvdata.kotlin.beatbox

import android.util.Log
import junit.framework.TestCase
import org.hamcrest.CoreMatchers.`is`
import org.hamcrest.MatcherAssert.assertThat
import org.junit.Before
import org.junit.Test
import org.mockito.Mockito.mock
import org.mockito.Mockito.verify

private const val TAG = "SoundViewModelTest"

class SoundViewModelTest {

    private lateinit var beatBox: BeatBox
    private lateinit var sound: Sound
    private lateinit var subject: SoundViewModel

    @Before
    fun setUp() {
        beatBox = mock(BeatBox::class.java)
        sound = Sound("assetPath")
        subject = SoundViewModel(beatBox)
        subject.sound = sound
        Log.d(TAG, "SoundViewModelTest::setUp() called!")
    }

    @Test
    fun exposesSoundNameAsTitle() {
        assertThat(subject.title, `is`(sound.name))
    }

    @Test
    fun callsBeatBoxPlayOnButtonClicked() {
        subject.onButtonClick()

        verify(beatBox).play(sound)
    }

    fun testSetSound() {}

    fun testGetTitle() {}

    fun testOnButtonClick() {}
}